Interrupt handling corrected

+++ EE209/src/lib/timer_driver.h
@@ -15,8 +15,8 @@
 void timer_init();
 void resetTimer();
 uint16_t getTimerCount();
-uint16_t calculateTime(uint16_t scale);
-uint16_t get_time();
+uint32_t calculateTime(uint16_t scale);
+uint32_t get_time();
 
 
 #endif /* TIMER_DRIVER_H_ */
